backend/requirements.txt
flask
flask-cors
psycopg2-binary
yfinance
pandas
requests
curl_cffi
python-dotenv
beautifulsoup4
firebase-admin
aiohttp
pydantic
pyyaml
openai
trafilatura
feedparser
tenacity
